添加了attitude文件中的函数，并把增稳的函数声明在了copter.h文件中，编译通过

+++ ahrs_DCM.h
@@ -64,9 +64,9 @@ public:
 
 	float		get_health(void);
 
-	long		roll_sensor;					// Degrees * 100
-	long		pitch_sensor;					// Degrees * 100
-	long		yaw_sensor;						// Degrees * 100
+	int32_t		roll_sensor;					// Degrees * 100
+	int32_t		pitch_sensor;					// Degrees * 100
+	int32_t		yaw_sensor;						// Degrees * 100
 
 	float		roll;							// Radians
 	float		pitch;							// Radians
